moves apply_array_uncertainty to CostMethods class
